Start experimenting with segregating compose files.
The project name is something I should have used to begin with. It'll be a nice
way of grouping containers that all have a common volume prefix.

However, now that I already have a number of containers that use the default
containers name, it will be hard to move them to something else.

For each environment I should probably

    have separate files per app deployment
    for example the atuin deployment would have its own compose
    and the name of the compose becomes the prefix for the name of the volumes.
    i'm not sure how this would work for shared volumes.