Start kicking the tires on docker-swarm.
The goal with this is going to be to get an atuin deployment working with the
service being able to communicate with the db.

I've got the env thing worked out, but it's hacky and i'll want to use secrets
instead. Currently atuin cannot resolve the name atuin-db, which is the next
thing to figure out.